fix: delete modal reference when removing a course
